move sendToRollbar to utils
