+++ app/actions/utils.js
@@ -21,7 +21,7 @@ import sundial from 'sundial';
 
 import ErrorMessages from '../constants/errorMessages';
 import * as syncActions from './sync';
-import { sendToRollbar } from './async';
+import rollbar from '../../app/utils/rollbar';
 
 const isBrowser = typeof window !== 'undefined';
 // eslint-disable-next-line no-console
@@ -134,6 +134,29 @@ export function mergeProfileUpdates(profile, updates){
   return _.mergeWith(profile, updates, (original, update) => {
     if (_.isArray(original)) {
       return update;
+    }
+  });
+}
+
+export function sendToRollbar(err, props) {
+  return new Promise((resolve) => {
+    if (rollbar) {
+      const extra = {};
+      if (_.get(props, 'data.blobId', false)) {
+        _.assign(extra, { blobId: props.data.blobId });
+      }
+      
+      rollbar.error(err, extra, (err, data) => {
+        if (err) {
+          console.log('Error while reporting error to Rollbar:', err);
+        } else {
+          console.log(`Rollbar UUID: ${data.result.uuid}`);
+          props.uuid = data.result.uuid;
+        }
+        resolve(props);
+      });
+    } else {
+      resolve(props);
     }
   });
 }
